rename eth2 pkg to pyspec, per request of hww
