+++ py_tests/phase0/block_processing/test_process_block_header.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 import pytest
 
 
-from eth2.phase0.spec import (
+from pyspec.phase0.spec import (
     get_beacon_proposer_index,
     cache_state,
     advance_slot,
