+++ py_tests/phase0/block_processing/test_process_proposer_slashing.py
@@ -1,8 +1,8 @@
 from copy import deepcopy
 import pytest
 
-import eth2.phase0.spec as spec
-from eth2.phase0.spec import (
+import pyspec.phase0.spec as spec
+from pyspec.phase0.spec import (
     get_balance,
     get_current_epoch,
     process_proposer_slashing,
