+++ py_tests/phase0/block_processing/test_voluntary_exit.py
@@ -1,9 +1,9 @@
 from copy import deepcopy
 import pytest
 
-import eth2.phase0.spec as spec
+import pyspec.phase0.spec as spec
 
-from eth2.phase0.spec import (
+from pyspec.phase0.spec import (
     get_active_validator_indices,
     get_current_epoch,
     process_voluntary_exit,
