+++ py_tests/phase0/helpers.py
@@ -2,9 +2,9 @@ from copy import deepcopy
 
 from py_ecc import bls
 
-import eth2.phase0.spec as spec
-from eth2.utils.minimal_ssz import signed_root
-from eth2.phase0.spec import (
+import pyspec.phase0.spec as spec
+from pyspec.utils.minimal_ssz import signed_root
+from pyspec.phase0.spec import (
     # constants
     EMPTY_SIGNATURE,
     ZERO_HASH,
@@ -33,7 +33,7 @@ from eth2.phase0.spec import (
     verify_merkle_branch,
     hash,
 )
-from eth2.utils.merkle_minimal import (
+from pyspec.utils.merkle_minimal import (
     calc_merkle_tree_from_leaves,
     get_merkle_proof,
     get_merkle_root,
