+++ test_libs/pyspec/pyspec/utils/minimal_ssz.py
